+++ hypothesis-python/src/hypothesis/core.py
@@ -50,7 +50,7 @@ from hypothesis._settings import (
     settings as Settings,
 )
 from hypothesis.control import BuildContext
-from hypothesis.database import ir_from_bytes, ir_to_bytes
+from hypothesis.database import choices_from_bytes, choices_to_bytes
 from hypothesis.errors import (
     BackendCannotProceed,
     DeadlineExceeded,
@@ -77,17 +77,16 @@ from hypothesis.internal.compat import (
     int_from_bytes,
 )
 from hypothesis.internal.conjecture.choice import ChoiceT
-from hypothesis.internal.conjecture.data import (
-    ConjectureData,
-    PrimitiveProvider,
-    Status,
-)
+from hypothesis.internal.conjecture.data import ConjectureData, Status
 from hypothesis.internal.conjecture.engine import BUFFER_SIZE, ConjectureRunner
 from hypothesis.internal.conjecture.junkdrawer import (
     ensure_free_stackframes,
     gc_cumulative_time,
 )
-from hypothesis.internal.conjecture.providers import BytestringProvider
+from hypothesis.internal.conjecture.providers import (
+    BytestringProvider,
+    PrimitiveProvider,
+)
 from hypothesis.internal.conjecture.shrinker import sort_key
 from hypothesis.internal.entropy import deterministic_PRNG
 from hypothesis.internal.escalation import (
@@ -323,7 +322,7 @@ def reproduce_failure(version: str, blob: bytes) -> Callable[[TestFunc], TestFun
 
 
 def encode_failure(choices):
-    blob = ir_to_bytes(choices)
+    blob = choices_to_bytes(choices)
     compressed = zlib.compress(blob)
     if len(compressed) < len(blob):
         blob = b"\1" + compressed
@@ -353,7 +352,7 @@ def decode_failure(blob: bytes) -> Sequence[ChoiceT]:
             f"Could not decode blob {blob!r}: Invalid start byte {prefix!r}"
         )
 
-    choices = ir_from_bytes(decoded)
+    choices = choices_from_bytes(decoded)
     if choices is None:
         raise InvalidArgument(f"Invalid serialized choice sequence for blob {blob!r}")
 
@@ -1065,9 +1064,7 @@ class StateForActualGivenExecution:
             for io in err._interesting_origins
             if io in self._runner.interesting_examples
         ]
-        exceptions = [
-            ie.extra_information._expected_exception for ie in interesting_examples
-        ]
+        exceptions = [ie.expected_exception for ie in interesting_examples]
         exceptions.append(context)  # the immediate exception
         return FlakyFailure(err.reason, exceptions)
 
@@ -1153,10 +1150,10 @@ class StateForActualGivenExecution:
                 # engine that this test run was interesting. This is the normal
                 # path for test runs that fail.
                 tb = get_trimmed_traceback()
-                info = data.extra_information
-                info._expected_traceback = format_exception(e, tb)  # type: ignore
-                info._expected_exception = e  # type: ignore
-                verbose_report(info._expected_traceback)  # type: ignore
+                data.expected_traceback = format_exception(e, tb)
+                data.expected_exception = e
+                assert data.expected_traceback is not None  # for mypy
+                verbose_report(data.expected_traceback)
 
                 self.failed_normally = True
 
@@ -1335,7 +1332,6 @@ class StateForActualGivenExecution:
 
         explanations = explanatory_lines(self.explain_traces, self.settings)
         for falsifying_example in self.falsifying_examples:
-            info = falsifying_example.extra_information
             fragments = []
 
             ran_example = runner.new_conjecture_data_ir(
@@ -1344,8 +1340,8 @@ class StateForActualGivenExecution:
             ran_example.slice_comments = falsifying_example.slice_comments
             tb = None
             origin = None
-            assert info is not None
-            assert info._expected_exception is not None
+            assert falsifying_example.expected_exception is not None
+            assert falsifying_example.expected_traceback is not None
             try:
                 with with_reporter(fragments.append):
                     self.execute_once(
@@ -1353,8 +1349,8 @@ class StateForActualGivenExecution:
                         print_example=not self.is_find,
                         is_final=True,
                         expected_failure=(
-                            info._expected_exception,
-                            info._expected_traceback,
+                            falsifying_example.expected_exception,
+                            falsifying_example.expected_traceback,
                         ),
                     )
             except StopTest as e:
@@ -1370,7 +1366,8 @@ class StateForActualGivenExecution:
                     "Inconsistent results: An example failed on the "
                     "first run but now succeeds (or fails with another "
                     "error, or is for some reason not runnable).",
-                    [info._expected_exception or e],  # (note: e is a BaseException)
+                    # (note: e is a BaseException)
+                    [falsifying_example.expected_exception or e],
                 )
                 errors_to_report.append((fragments, err))
             except UnsatisfiedAssumption as e:  # pragma: no cover  # ironically flaky
@@ -1864,7 +1861,7 @@ def given(
             minimal_failures: dict = {}
 
             def fuzz_one_input(
-                buffer: Union[bytes, bytearray, memoryview, BinaryIO]
+                buffer: Union[bytes, bytearray, memoryview, BinaryIO],
             ) -> Optional[bytes]:
                 # This inner part is all that the fuzzer will actually run,
                 # so we keep it as small and as fast as possible.
@@ -1885,7 +1882,9 @@ def given(
                     if settings.database is not None and (
                         known is None or sort_key(data.nodes) <= sort_key(known)
                     ):
-                        settings.database.save(database_key, ir_to_bytes(data.choices))
+                        settings.database.save(
+                            database_key, choices_to_bytes(data.choices)
+                        )
                         minimal_failures[data.interesting_origin] = data.nodes
                     raise
                 assert isinstance(data.provider, BytestringProvider)